Reduce duplicated config to single
